1 урок

+++ src/lesson1/task1/Simple.kt
@@ -53,8 +53,8 @@ fun main(args: Array<String>) {
  */
 fun seconds(hours: Int, minutes: Int, seconds: Int) :Int {
     val a:Int = hours * 3600
-        val b:Int = minutes * 60
-            val c:Int = seconds
+    val b:Int = minutes * 60
+    val c:Int = seconds
     return a + b + c
 }
 /**
@@ -66,8 +66,8 @@ fun seconds(hours: Int, minutes: Int, seconds: Int) :Int {
  */
 fun lengthInMeters(sagenes: Int, arshins: Int, vershoks: Int): Double {
     val a:Int = sagenes * 48
-        val b:Int = arshins * 16
-            val c:Int = vershoks
+    val b:Int = arshins * 16
+    val c:Int = vershoks
     return (a + b +c ) * 4.445 /100
 }
 
@@ -88,7 +88,7 @@ fun angleInRadian(grad: Int, min: Int, sec: Int): Double {
  * Например, расстояние между (3, 0) и (0, 4) равно 5
  */
 fun trackLength(x1: Double, y1: Double, x2: Double, y2: Double): Double {
-    return sqrt(sqr(x1 - x2).toDouble() + sqr(y1 - y2))
+    return sqrt(sqr(x1 - x2) + sqr(y1 - y2))
 
 }
 
@@ -111,7 +111,7 @@ fun thirdDigit(number: Int): Int {
  */
 fun travelMinutes(hoursDepart: Int, minutesDepart: Int, hoursArrive: Int, minutesArrive: Int): Int {
     val a:Int = (hoursArrive  * 60 + minutesArrive)
-     val b:Int = (hoursDepart * 60 + minutesDepart)
+    val b:Int = (hoursDepart * 60 + minutesDepart)
     return a - b
 }
 
@@ -123,10 +123,10 @@ fun travelMinutes(hoursDepart: Int, minutesDepart: Int, hoursArrive: Int, minute
  * Например, 100 рублей под 10% годовых превратятся в 133.1 рубля
  */
 fun accountInThreeYears(initial: Int, percent: Int): Double {
-   val firstYear:Double = initial * (1 + percent / 100.0)
-        val secondYear:Double = firstYear * (1 + percent / 100.0)
+    val firstYear:Double = initial * (1 + percent / 100.0)
+    val secondYear:Double = firstYear * (1 + percent / 100.0)
     val thirdYear:Double = secondYear * (1 + percent / 100.0)
-        return thirdYear
+    return thirdYear
 }
 
 /**
@@ -137,7 +137,7 @@ fun accountInThreeYears(initial: Int, percent: Int): Double {
  */
 fun numberRevert(number: Int): Int {
     val firstNumeral:Int = number % 10 * 100
-        val secondNumeral:Int =  number % 100 / 10 * 10
-            val thirdNumeral:Int = number % 1000 / 100
+    val secondNumeral:Int =  number % 100 / 10 * 10
+    val thirdNumeral:Int = number % 1000 / 100
     return firstNumeral + secondNumeral + thirdNumeral
 }
